jsonformat generator
